ビルトインコマンドの実行を分離
